CampusOS/requirements.txt
# CampusOS — Education-to-Employment Operating System
# Core Django + DRF
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Celery + Redis
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Storage
Pillow==10.3.0
django-storages==1.14.3
boto3==1.34.69

# Auth + Security
cryptography==42.0.5
PyJWT==2.8.0
django-ratelimit==4.1.0

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0
phonenumbers==8.13.35

# AI / ML
openai==1.23.2
tiktoken==0.6.0

# Email
sendgrid==6.11.0

# SMS / WhatsApp
twilio==9.0.5

# PDF generation
weasyprint==62.1
reportlab==4.1.0

# Export / Import
openpyxl==3.1.2
pandas==2.2.1

# Search
elasticsearch==8.13.0

# Monitoring
sentry-sdk==1.43.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Code quality
flake8==7.0.0
black==24.4.2
isort==5.13.2
mypy==1.9.0
django-stubs==5.0.0

# ConnectOS shared platform layer
connectos-platform @ file:../../ConnectOS/platform
